vm_01/requirements.txt
Flask
flask-cors
google-cloud-storage
Pillow
line-bot-sdk
uwsgi
pymongo
gunicorn